+++ options.h
@@ -6,7 +6,7 @@
 #define FOOD_ORDERING_OPTIONS_H
 
 void printMenu(int nrOfTypes, char **types);
-void printMenuWithPrices(int id, int nrOfTypes, char **types, int *prices);
+void printMenuWithPrices(int id, int nrOfTypes, char **types, double *prices);
 int makeChoice(int *state, int nrOfTypes, int stateModifier);
 void getAdditionalInfo(int *state, char additionalInfo[]);
